Home Page Updates
Adjusted the absolute pos. on a lot of things on the page to make the top image fit correctly.

Added a new "Best Sellers" section.

Made a WatchCard.tsx component and added 3 to the best sellers section.

Adjusted scroll bar colors.

Added a Home.module.css file that has a glow effect that is being used on the home page.

Removed unneeded code from the global css file.