Updates
Added a Review Slider Component.

Customized the review slider component.

Positioned an image behind the review slider and changed its sizes ( still needs to be adjusted ).

Added a copy of the about us section to get a better visual on the height of the review slider section.

Added text above the review slider section.