Updates
Fixed Hydration Error by removing the paragraph prop from the About Us Card and directly coding the paragraph in the component itself.

Added a Footer component.

Dropped the hover line on the navbar links by 1.

Made the images in the Watch Card component bigger as well as the width of the card itself.

Adjusted the width of the buttons on Watch Card to fit the size adjustments.

Made the price text large and font bold.

Corrected the hover duration times on all buttons to be more responsive.