Updates
Added a Custom 404 Page.

Built a Filter Component on the Shop Page that filters the watch cards prices.

Added a Shopping Cart Icon from React Icons to the NavBar for future checkout page.

Made the "View Shop" button link to the Shop page.

Changed the price text color on the watch cards to green.