Updates
Made Detail Page dynamic and showcase the specific watch card the user clicked on.

Fixed the detail information for watch cards on home page.

Fixed the detail information for watch cards on shop page.

Fixed issue with the star rating on the detail page not showing correct number of stars based on its watch card.

Fixed image display for images on Detail Page.