Updates
Added an Item Added to Cart overlay when the user clicks add to cart.

Added a remove item from cart reducer function to the cart slice file and adjusted the cart item component to use this function so the user can remove an item from the cart.

Made the shopping cart icon in the NavBar show the count of how many items the user has in their cart.

Adjusted the totals of the cart total and cart item total so it correctly displays the total price ( still needs work )