Updates
Added dynamic text and link button to cart page if user has no items in cart.

Added the "add to cart" functionality to the details description button so when the user clicks it the item is added to the cart page as a cart item.

Added "item added to cart" overlay for the button on the details page.

Removed unneeded code from details description component.

Added an empty component to be worked on later.

Other small changes.